pyproject.toml
[project]
name = "rider-community-backend"
version = "0.1.0"
description = "Rider community backend (Django + FastAPI)"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.13",
    "django-cors-headers>=4.9.0",
    "fastapi>=0.136.1",
    "pillow>=12.2.0",
    "pydantic[email]>=2.13.3",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "uvicorn[standard]>=0.46.0",
]
